# Merging with 32bit precision, reduced memory usage and improved performance