--- a/burstphoto/align/align.metal
+++ b/burstphoto/align/align.metal
@@ -6,6 +6,7 @@ using namespace metal;
 kernel void avg_pool(texture2d<float, access::read> in_texture [[texture(0)]],
                      texture2d<float, access::write> out_texture [[texture(1)]],
                      constant int& scale [[buffer(0)]],
+                     constant float& black_level [[buffer(1)]],
                      uint2 gid [[thread_position_in_grid]]) {
     
     float out_pixel = 0;
@@ -16,7 +17,7 @@ kernel void avg_pool(texture2d<float, access::read> in_texture [[texture(0)]],
         for (int dy = 0; dy < scale; dy++) {
             int x = x0 + dx;
             int y = y0 + dy;
-            out_pixel += in_texture.read(uint2(x, y)).r;
+            out_pixel += (in_texture.read(uint2(x, y)).r - black_level);
         }
     }
     
@@ -28,9 +29,10 @@ kernel void avg_pool(texture2d<float, access::read> in_texture [[texture(0)]],
 kernel void avg_pool_normalization(texture2d<float, access::read> in_texture [[texture(0)]],
                                    texture2d<float, access::write> out_texture [[texture(1)]],
                                    constant int& scale [[buffer(0)]],
-                                   constant float& factor_red [[buffer(1)]],
-                                   constant float& factor_green [[buffer(2)]],
-                                   constant float& factor_blue [[buffer(3)]],
+                                   constant float& black_level [[buffer(1)]],
+                                   constant float& factor_red [[buffer(2)]],
+                                   constant float& factor_green [[buffer(3)]],
+                                   constant float& factor_blue [[buffer(4)]],
                                    uint2 gid [[thread_position_in_grid]]) {
     
     float out_pixel = 0;
@@ -44,7 +46,7 @@ kernel void avg_pool_normalization(texture2d<float, access::read> in_texture [[t
         for (int dy = 0; dy < scale; dy++) {
             int x = x0 + dx;
             int y = y0 + dy;
-            out_pixel += (mean_factor/norm_factors[dy*scale+dx]*in_texture.read(uint2(x, y)).r);
+            out_pixel += (mean_factor/norm_factors[dy*scale+dx]*in_texture.read(uint2(x, y)).r - black_level);
         }
     }
 
@@ -63,8 +65,7 @@ kernel void compute_tile_differences(texture2d<float, access::read> ref_texture 
                                      constant int& downscale_factor [[buffer(0)]],
                                      constant int& tile_size [[buffer(1)]],
                                      constant int& search_dist [[buffer(2)]],
-                                     constant float& black_level [[buffer(3)]],
-                                     constant int& weight_ssd [[buffer(4)]],
+                                     constant int& weight_ssd [[buffer(3)]],
                                      uint3 gid [[thread_position_in_grid]]) {
     
     // load args
@@ -123,8 +124,7 @@ kernel void compute_tile_differences25(texture2d<half, access::read> ref_texture
                                        constant int& downscale_factor [[buffer(0)]],
                                        constant int& tile_size [[buffer(1)]],
                                        constant int& search_dist [[buffer(2)]],
-                                       constant float& black_level [[buffer(3)]],
-                                       constant int& weight_ssd [[buffer(4)]],
+                                       constant int& weight_ssd [[buffer(3)]],
                                        uint2 gid [[thread_position_in_grid]]) {
     
     // load args
@@ -215,7 +215,7 @@ kernel void compute_tile_differences25(texture2d<half, access::read> ref_texture
     
     // store tile differences in texture
     for (int i = 0; i < 25; i++) {
-        tile_diff.write(diff[i], uint3(gid.x, gid.y, i));
+        tile_diff.write(diff[i], uint3(i, gid.x, gid.y));
     }
 }
 
@@ -230,15 +230,12 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
                                                 constant int& downscale_factor [[buffer(0)]],
                                                 constant int& tile_size [[buffer(1)]],
                                                 constant int& search_dist [[buffer(2)]],
-                                                constant float& black_level [[buffer(3)]],
-                                                constant int& weight_ssd [[buffer(4)]],
+                                                constant int& weight_ssd [[buffer(3)]],
                                                 uint2 gid [[thread_position_in_grid]]) {
     
     // load args
     int const texture_width = ref_texture.get_width();
     int const texture_height = ref_texture.get_height();
-    
-    half const black_level_float16 = half(black_level);
      
     int ref_tile_x, ref_tile_y, comp_tile_x, comp_tile_y, tmp_index, dx_i, dy_i;
     
@@ -275,7 +272,7 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
             if ((comp_tile_x < 0) || (comp_tile_y < 0) || (comp_tile_x >= texture_width) || (comp_tile_y >= texture_height)) {
                 tmp_comp[tmp_index] = FLOAT16_MAX_VAL;
             } else {
-                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * (comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r - black_level_float16));
+                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r);
             }
         }
     }
@@ -296,7 +293,7 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
             if ((comp_tile_x < 0) || (comp_tile_y < 0) || (comp_tile_x >= texture_width) || (comp_tile_y >= texture_height)) {
                 tmp_comp[tmp_index] = FLOAT16_MAX_VAL;
             } else {
-                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * (comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r - black_level_float16));
+                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r);
             }
         }
         
@@ -306,7 +303,7 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
             ref_tile_x = x0 + dx;
             ref_tile_y = y0 + dy;
             
-            tmp_ref = max(FLOAT16_ZERO_VAL, ref_texture.read(uint2(ref_tile_x, ref_tile_y)).r - black_level_float16);
+            tmp_ref = max(FLOAT16_ZERO_VAL, ref_texture.read(uint2(ref_tile_x, ref_tile_y)).r);
               
             // loop over 25 test displacements
             for (int i = 0; i < 25; i++) {
@@ -348,7 +345,7 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
             if ((comp_tile_x < 0) || (comp_tile_y < 0) || (comp_tile_x >= texture_width) || (comp_tile_y >= texture_height)) {
                 tmp_comp[tmp_index] = FLOAT16_MIN_VAL;
             } else {
-                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * (comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r - black_level_float16));
+                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r);
             }
         }
     }
@@ -369,7 +366,7 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
             if ((comp_tile_x < 0) || (comp_tile_y < 0) || (comp_tile_x >= texture_width) || (comp_tile_y >= texture_height)) {
                 tmp_comp[tmp_index] = FLOAT16_MIN_VAL;
             } else {
-                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * (comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r - black_level_float16));
+                tmp_comp[tmp_index] = max(FLOAT16_ZERO_VAL, FLOAT16_05_VAL * comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r);
             }
         }
         
@@ -379,7 +376,7 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
             ref_tile_x = x0 + dx;
             ref_tile_y = y0 + dy;
             
-            tmp_ref = max(FLOAT16_ZERO_VAL, ref_texture.read(uint2(ref_tile_x, ref_tile_y)).r - black_level_float16);
+            tmp_ref = max(FLOAT16_ZERO_VAL, ref_texture.read(uint2(ref_tile_x, ref_tile_y)).r);
               
             // loop over 25 test displacements
             for (int i = 0; i < 25; i++) {
@@ -400,7 +397,7 @@ kernel void compute_tile_differences_exposure25(texture2d<half, access::read> re
     
     // store tile differences in texture
     for (int i = 0; i < 25; i++) {
-        tile_diff.write(diff[i], uint3(gid.x, gid.y, i));
+        tile_diff.write(diff[i], uint3(i, gid.x, gid.y));
     }
 }
 
@@ -417,16 +414,13 @@ kernel void correct_upsampling_error(texture2d<half, access::read> ref_texture [
                                      constant int& tile_size [[buffer(1)]],
                                      constant int& n_tiles_x [[buffer(2)]],
                                      constant int& n_tiles_y [[buffer(3)]],
-                                     constant float& black_level [[buffer(4)]],
-                                     constant int& uniform_exposure [[buffer(5)]],
-                                     constant int& weight_ssd [[buffer(6)]],
+                                     constant int& uniform_exposure [[buffer(4)]],
+                                     constant int& weight_ssd [[buffer(5)]],
                                      uint2 gid [[thread_position_in_grid]]) {
     
     // load args
     int const texture_width = ref_texture.get_width();
     int const texture_height = ref_texture.get_height();
-    
-    half const black_level_float16 = half(black_level);
     
     // initialize some variables
     int comp_tile_x, comp_tile_y, tmp_tile_x, tmp_tile_y, weight_outside;
@@ -467,7 +461,7 @@ kernel void correct_upsampling_error(texture2d<half, access::read> ref_texture [
             
             // copy 64/tile_size rows into temp vector
             for (int i = 0; i < 64; i++) {
-                tmp_ref[i] = max(FLOAT16_ZERO_VAL, ref_texture.read(uint2(x0+(i % tile_size), y0+dy+int(i / tile_size))).r - black_level_float16);
+                tmp_ref[i] = max(FLOAT16_ZERO_VAL, ref_texture.read(uint2(x0+(i % tile_size), y0+dy+int(i / tile_size))).r);
             }
             
             // loop over three candidates
@@ -485,7 +479,7 @@ kernel void correct_upsampling_error(texture2d<half, access::read> ref_texture [
                     if ((comp_tile_x >= 0) && (comp_tile_y >= 0) && (comp_tile_x < texture_width) && (comp_tile_y < texture_height)) {
                         
                         sum_u[c] += tmp_ref[i];
-                        sum_v[c] += max(FLOAT16_ZERO_VAL, comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r - black_level_float16);
+                        sum_v[c] += max(FLOAT16_ZERO_VAL, comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r);
                     }
                 }
             }
@@ -502,7 +496,7 @@ kernel void correct_upsampling_error(texture2d<half, access::read> ref_texture [
         
         // copy 64/tile_size rows into temp vector
         for (int i = 0; i < 64; i++) {
-            tmp_ref[i] = ref_texture.read(uint2(x0+(i % tile_size), y0+dy+int(i / tile_size))).r - black_level_float16;
+            tmp_ref[i] = ref_texture.read(uint2(x0+(i % tile_size), y0+dy+int(i / tile_size))).r;
         }
         
         // loop over three candidates
@@ -521,7 +515,7 @@ kernel void correct_upsampling_error(texture2d<half, access::read> ref_texture [
                 weight_outside = clamp(texture_width-comp_tile_x-1, -1, 0) + clamp(texture_height-comp_tile_y-1, -1, 0) + clamp(comp_tile_x, -1, 0) + clamp(comp_tile_y, -1, 0);
                 weight_outside = -max(-1, weight_outside);
                 
-                diff_abs = abs(tmp_ref[i] - (1-weight_outside)*ratio[c]*(comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r - black_level_float16) - weight_outside*2*FLOAT16_MIN_VAL);
+                diff_abs = abs(tmp_ref[i] - (1-weight_outside)*ratio[c]*(comp_texture.read(uint2(comp_tile_x, comp_tile_y)).r) - weight_outside*2*FLOAT16_MIN_VAL);
                 
                 // add difference to corresponding combination
                 diff[c] += (1-weight_ssd)*diff_abs + weight_ssd*diff_abs*diff_abs;
@@ -557,7 +551,7 @@ kernel void find_best_tile_alignment(texture3d<float, access::read> tile_diff [[
     int min_diff_idx = 0;
     
     for (int i = 0; i < n_pos_2d; i++) {
-        current_diff = tile_diff.read(uint3(gid.x, gid.y, i)).r;
+        current_diff = tile_diff.read(uint3(i, gid.x, gid.y)).r;
         if (current_diff < min_diff_val) {
             min_diff_val = current_diff;
             min_diff_idx = i;
@@ -577,8 +571,8 @@ kernel void find_best_tile_alignment(texture3d<float, access::read> tile_diff [[
 }
 
 
-kernel void warp_texture_bayer(texture2d<half, access::read> in_texture [[texture(0)]],
-                               texture2d<half, access::write> out_texture [[texture(1)]],
+kernel void warp_texture_bayer(texture2d<float, access::read> in_texture [[texture(0)]],
+                               texture2d<float, access::write> out_texture [[texture(1)]],
                                texture2d<int, access::read> prev_alignment [[texture(2)]],
                                constant int& downscale_factor [[buffer(0)]],
                                constant int& half_tile_size [[buffer(1)]],