… burstphoto/Assets.xcassets/camera_icon.symbolset/Contents.json
… burstphoto/Assets.xcassets/camera_icon.symbolset/camera.svg
… burstphoto/Assets.xcassets/stack_icon.symbolset/Contents.json
… burstphoto/Assets.xcassets/stack_icon.symbolset/square.3.layers.3d.svg
… burstphoto/exposure/exposure.swift
… burstphoto/merge/frequency.swift
… burstphoto/merge/spatial.swift
… burstphoto/texture/texture.metal
… burstphoto/texture/texture.swift