@chris-rank 2023-10-08
Yes, this could be a version 1.6 release. But we should align with Alex first as he has also an open PR. While I already tested the code a lot, some review or testing of someone else would be great.