@Alex-Vasile 2023-10-08
Re: dealing with the black level. I was thinking about this, but how come we don't subtract the black level once as part of reading the DNG from file into the first metal texture. Beyond that we shouldn't need it anymore.