@chris-rank 2023-10-08
I had a look into the description in Apple's documentation and did not find a way to explicitly trigger lossless texture compression. It might not even be supported properly for the macOS version 11 we currently build the app for. As the peak memory usage is much lower now, I thought that it may not be needed anymore. My interpretation of your issue was that some evaluation of that feature shall be done.