@Alex-Vasile 2023-10-08
> I had a look into the description in Apple's documentation and did not find a way to explicitly trigger lossless texture compression. It might not even be supported properly for the macOS version 11 we currently build the app for. As the peak memory usage is much lower now, I thought that it may not be needed anymore. My interpretation of your issue was that some evaluation of that feature shall be done.

The peak is much better now, but the peak will still slowly increase until the limit set for the cache. The compression was for trying to increase the amount of photos that could be cached without increase the memory usage.