@chris-rank 2023-10-09
If you prefer, I can remove the #32 from this PR. You are right, a compression would enable more frames on the cache.