@Alex-Vasile 2023-10-09
> If you prefer, I can remove the #32 from this PR. You are right, a compression would enable more frames on the cache.

I think you should. We can close that issue separately if we don't find a solution, but I don't believe that this PR (good as it is) closes it.