@Alex-Vasile 2023-10-13
Overall, could you please provide more information about how this PR reduces peak and cumulative memory usage?